+++ src/main/com/tutorial/projectwithMaven/Customer.java
@@ -1,0 +1,70 @@
+package customerRegistration.coreJava.src.main.com.tutorial.projectwithMaven;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class Customer {
+    @Id
+    @GeneratedValue(strategy= GenerationType.IDENTITY)
+    int id;
+    String name;
+    int age;
+    String mobile;
+
+
+    Customer() {
+
+    }
+
+    Customer(int id, String name,int age, String mobile) {
+        this.id = id;
+        this.name = name;
+        this.age = age;
+        this.mobile = mobile;
+
+    }
+    public int getId()
+    {
+        return id;
+    }
+    public String getName() {
+        return name;
+    }
+
+    public int getAge() {
+        return age;
+    }
+
+    public String getMobileNumber() {
+        return mobile;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public void setAge(int age) {
+        this.age = age;
+    }
+
+    public void setMobile(String mobile) {
+        this.mobile = mobile;
+    }
+
+    @Override
+    public String toString() {
+        return "Customer{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", age=" + age +
+                ", mobile='" + mobile + '\'' +
+                '}';
+    }
+}
